requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
passlib
yfinance
pandas
pandas-ta
anthropic
python-dotenv
httpx
google-generativeai
